dir improvements: move `spectest` into `cl/spectest`
