+++ cl/spectest/consensus_tests/bls.go
@@ -20,7 +20,7 @@ import (
 	"io/fs"
 	"testing"
 
-	"github.com/erigontech/erigon/spectest"
+	"github.com/erigontech/erigon/cl/spectest/spectest"
 )
 
 type BlsAggregateVerify struct {
