+++ cl/spectest/consensus_tests/light_client.go
@@ -20,12 +20,13 @@ import (
 	"io/fs"
 	"testing"
 
+	"github.com/stretchr/testify/require"
+
 	"github.com/erigontech/erigon/cl/clparams"
 	"github.com/erigontech/erigon/cl/cltypes"
 	"github.com/erigontech/erigon/cl/phase1/core/state"
+	"github.com/erigontech/erigon/cl/spectest/spectest"
 	"github.com/erigontech/erigon/common"
-	"github.com/erigontech/erigon/spectest"
-	"github.com/stretchr/testify/require"
 )
 
 type LcBranch struct {
