+++ cl/spectest/spectest/case.go
