+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_aggregate_pubkeys/small/eth_aggregate_pubkeys_valid_e235e92e3a313f43/data.yaml
