+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_aggregate_pubkeys/small/eth_aggregate_pubkeys_x40_pubkey/data.yaml
