+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_aggregate_pubkeys/small/eth_aggregate_pubkeys_zero_pubkey/data.yaml
