+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_fast_aggregate_verify/small/eth_fast_aggregate_verify_extra_pubkey_5a38e6b4017fe4dd/data.yaml
