+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_fast_aggregate_verify/small/eth_fast_aggregate_verify_extra_pubkey_a698ea45b109f303/data.yaml
