+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_fast_aggregate_verify/small/eth_fast_aggregate_verify_infinity_pubkey/data.yaml
