+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_fast_aggregate_verify/small/eth_fast_aggregate_verify_valid_3d7576f3c0e3570a/data.yaml
