+++ cl/spectest/spectest/data_faketest/tests/general/altair/bls/eth_fast_aggregate_verify/small/eth_fast_aggregate_verify_valid_5e745ad0c6199a6c/data.yaml
